+++ routes/questions.js
@@ -3,8 +3,11 @@ const router = express.Router();
 const models = require("../models");
 
 router.get('/', (req, res) => {
-	models.questions.findAll().then(data => res.json(data));
-})
+	models.questions.findAll()
+	.then(data => {
+		res.status(200).json(data)
+	});
+});
 
 router.get('/:id(\\d+)', (req, res) => {
 	models.questions.findAll({
@@ -12,8 +15,10 @@ router.get('/:id(\\d+)', (req, res) => {
 			id : req.params.id
 		}
 	})
-	.then(data => res.json(data));
-})
+	.then(data => {
+		res.status(200).json(data)
+	});
+});
 
 router.post('/', (req, res) => {
 	const data = req.body;
@@ -26,27 +31,45 @@ router.post('/', (req, res) => {
 		.catch(err => {
 			res.status(500).send('Cannot add Question')
 		});
-})
+});
 
 router.put('/:id(\\d+)', (req, res) => {
-	const data = req.body;
-	console.log(data);
-	models.questions.update(
-		data,
-		{ where : { id : req.params.id } }
-	);
-
-	res.sendStatus(200);
-})
-
-router.delete('/:id(\\d+)', (req, res) => {
-	models.questions.destroy({
-		where : {
-			id : req.params.id
+	models.questions.findById(req.params.id)
+	.then(questionsFound => {
+		if(questionsFound){
+			const data = req.body;
+			console.log(data);
+			models.questions.update(
+				data,
+				{ where : { id : req.params.id } }
+			)
+			.then(updatedQuestions => {
+				res.status(200).send(`Question updated at id : ${req.params.id }`);
+			});
+		}
+		else{
+			return res.status(404).send(`Question ${req.params.id} does not exist in DB`);
 		}
 	});
+});
 
-	res.sendStatus(200);
-})
+router.delete('/:id(\\d+)', (req, res) => {
+	models.questions.findById(req.params.id)
+	.then(questionsFound => {
+		if(questionsFound){
+			models.questions.destroy({
+				where : {
+					id : req.params.id
+				}
+			})
+			.then(updatedQuestions => {
+				res.status(200).send(`Question deleted at id : ${req.params.id }`);
+			})
+		}
+		else{
+			return res.status(404).send(`Question ${req.params.id} does not exist in DB`);
+		}
+	});
+});
 
 module.exports = router;
