+++ routes/users.js
@@ -4,8 +4,11 @@ const models = require('../models');
 const userCtrl = require('./log');
 
 router.get('/', (req, res) => {
-	models.users.findAll().then(data => res.json(data));
-})
+	models.users.findAll()
+	.then(data => {
+		res.status(200).json(data)
+	});
+});
 
 router.get('/:id(\\d+)', (req, res) => {
 	models.users.findAll({
@@ -13,32 +16,52 @@ router.get('/:id(\\d+)', (req, res) => {
 			id : req.params.id
 		}
 	})
-	.then(data => res.json(data));
-})
+	.then(data => {
+		res.sattus(200).json(data)
+	});
+});
 
 router.post('/register/', userCtrl.register);
 
 router.post('/login/', userCtrl.login);
 
 router.put('/:id(\\d+)', (req, res) => {
-	const data = req.body;
-	console.log(data);
-	models.users.update(
-		data,
-		{ where : { id : req.params.id } }
-	);
-
-	res.sendStatus(200);
-})
-
-router.delete('/:id(\\d+)', (req, res) => {
-	models.users.destroy({
-		where : {
-			id : req.params.id
+	models.users.findById(req.params.id)
+	.then(usersFound => {
+		if(usersFound){
+			const data = req.body;
+			console.log(data);
+			models.users.update(
+				data,
+				{ where : { id : req.params.id } }
+			)
+			.then(updateUser => {
+				res.status(200).send(`User updated at id : ${req.params.id }`);
+			});
+		}
+		else{
+			return res.status(404).send(`User ${req.params.id} does not exist in DB`);
 		}
 	});
+});
 
-	res.sendStatus(200);
-})
+router.delete('/:id(\\d+)', (req, res) => {
+	models.users.findById(req.params.id)
+	.then(usersFound => {
+		if(usersFound){
+			models.users.destroy({
+				where : {
+					id : req.params.id
+				}
+			})
+			.then(updateUser => {
+				res.status(200).send(`User deleted at id : ${req.params.id }`);
+			})
+		}
+		else{
+			return res.status(404).send(`User ${req.params.id} does not exist in DB`);
+		}
+	});
+});
 
 module.exports = router;
